# project-1/docs/swagger.yaml
definitions:
  controller.CommonError:
    properties:
      error:
        type: string
    type: object
  controller.CommonResponse:
    properties:
      message:
        type: string
    type: object
  dto.EmployeeCreateRequest:
    properties:
      age:
        type: integer
      email:
        type: string
      name:
        type: string
    type: object
  models.Employee:
    properties:
      age:
        type: integer
      email:
        type: string
      id:
        type: integer
      name:
        type: string
    type: object
info:
  contact:
    email: test@test.com
    name: OnlineShop API Service
    url: http://test.com
  title: OnlineShop API
paths:
  /employee/{id}:
    delete:
      description: Delete employee by ID or throw an error
      parameters:
      - description: employee id
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/controller.CommonResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/controller.CommonError'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/controller.CommonError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/controller.CommonError'
      summary: Delete employee by ID
      tags:
      - Employees
  /employees:
    get:
      description: Get all employees or throw an error
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.Employee'
            type: array
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/controller.CommonError'
      summary: Get all employees
      tags:
      - Employees
    post:
      description: Add new employee or throw an error
      parameters:
      - description: new employee info
        in: body
        name: request_body
        required: true
        schema:
          $ref: '#/definitions/dto.EmployeeCreateRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/controller.CommonResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/controller.CommonError'
        "422":
          description: Unprocessable Entity
          schema:
            $ref: '#/definitions/controller.CommonError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/controller.CommonError'
      summary: Add new employee
      tags:
      - Employees
  /employees/{id}:
    get:
      description: Get employee by ID or throw an error
      parameters:
      - description: id of employee
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Employee'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/controller.CommonError'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/controller.CommonError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/controller.CommonError'
      summary: Get employee by ID
      tags:
      - Employees
    put:
      description: Update employee by ID or throw an error
      parameters:
      - description: employee id
        in: path
        name: id
        required: true
        type: integer
      - description: employee info
        in: body
        name: request_body
        required: true
        schema:
          $ref: '#/definitions/dto.EmployeeCreateRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/controller.CommonResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/controller.CommonError'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/controller.CommonError'
        "422":
          description: Unprocessable Entity
          schema:
            $ref: '#/definitions/controller.CommonError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/controller.CommonError'
      summary: Update employee by ID
      tags:
      - Employees
swagger: "2.0"